requirements.txt
celery
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
argon2-cffi
python-jose